Implement autocompletion in linedit.fs
Add autocompletion to linedit.fs.  Completion is carried out when the
cursor is in the end of a word.  It looks for all accessible words (in
the search order) and complete the first compatible one.  If you press
more times TAB, you will get more coincidences.  Any other key
restarts the completion.