Fix bug in end of vocabularies handling
It was a hard nut to crack. Eulex did not boot on Bochs emulator
either _some_ hardware machines, but I carried off to find the bug
finally, and then it was not dificult to fix it.

It was located in the routine `find_word_1'. Trying to find a word in
an empty wordlist caused a crash because the checking for null word
was in the end of the routine.  Now, find_word_1 checks nullity in the
beginning.

DVP